+++ lib/tableficate/helper.rb
@@ -3,7 +3,11 @@ module Tableficate
     def table_for(rows, options = {})
       t = Tableficate::Table.new(self, rows, options, rows.tableficate_get_data)
       yield(t)
-      t.template.render(partial: Tableficate::Utils::template_path('table', t.theme), locals:  {table: t})
+      t.template.render(partial: Tableficate::Utils::template_path('table_for', t.theme), locals:  {table: t})
+    end
+
+    def tableficate_table_tag(table)
+      render partial: Tableficate::Utils::template_path('table', table.theme), locals: {table: table}
     end
 
     def tableficate_header_tag(column)